Change fp type to float

+++ examples/oneapi/dpc/source/logistic_regression/logistic_regression_dense_batch.cpp
@@ -63,11 +63,11 @@ void run(sycl::queue& q) {
     const auto optimizer_desc = dal::newton_cg::descriptor<>(1e-4, 100l);
 
     const auto log_reg_desc =
-        dal::logistic_regression::descriptor<double, method_t, task_t, optimizer_t>(true,
-                                                                                    0.0,
-                                                                                    1.0,
-                                                                                    2,
-                                                                                    optimizer_desc)
+        dal::logistic_regression::descriptor<float, method_t, task_t, optimizer_t>(true,
+                                                                                   0.0,
+                                                                                   1.0,
+                                                                                   2,
+                                                                                   optimizer_desc)
             .set_result_options(result_options::coefficients | result_options::intercept |
                                 result_options::iterations_number);
